Prevent warnings from type conversion

+++ IGraphics/IGraphicsStructs.h
@@ -248,9 +248,9 @@ struct IColor
   static IColor FromRGBf(float* rgbf)
   {
     int A = 255;
-    int R = rgbf[0] * 255;
-    int G = rgbf[1] * 255;
-    int B = rgbf[2] * 255;
+    int R = (int)(rgbf[0] * 255.f);
+    int G = (int)(rgbf[1] * 255.f);
+    int B = (int)(rgbf[2] * 255.f);
     
     return IColor(A, R, G, B);
   }
@@ -260,10 +260,10 @@ struct IColor
    * @return IColor A new IColor based on the input array */
   static IColor FromRGBAf(float* rgbaf)
   {
-    int R = rgbaf[0] * 255;
-    int G = rgbaf[1] * 255;
-    int B = rgbaf[2] * 255;
-    int A = rgbaf[3] * 255;
+    int R = (int)(rgbaf[0] * 255.f);
+    int G = (int)(rgbaf[1] * 255.f);
+    int B = (int)(rgbaf[2] * 255.f);
+    int A = (int)(rgbaf[3] * 255.f);
 
     return IColor(A, R, G, B);
   }
